fix(statistics): remove redundant pivot casts for static analysis
Drop unnecessary float casts in pivot helpers so Psalm no longer reports redundant-cast issues while preserving numeric behavior.

+++ src/Statistics/Application/Pivot/PivotTableBuilder.php
@@ -26,7 +26,7 @@ final class PivotTableBuilder
             $row = [];
             $rowTotal = 0.0;
             foreach ($colKeys as $colKey) {
-                $value = (float) ($index[$rowKey.'|'.$colKey] ?? 0.0);
+                $value = $index[$rowKey.'|'.$colKey] ?? 0.0;
                 $row[] = $value;
                 $rowTotal += $value;
             }
